pyproject.toml
[tool.poetry]
name = "post-manager"
version = "1.0"
description = "A post manager for Routine of Damak Banda"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "post_manager"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.5.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.4"
beautifulsoup4 = "^4.12.2"
aiocron = "^1.8"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
